// src/typeScript/interfaces/pokemon.ts
export interface Pokemon_Base {
  name:       string
  typeId:     string
  idPokemon:  string
}